requirements.txt
langgraph>=0.3.0
langchain>=0.0.193
openai>=0.27.0
chromadb>=0.3.31
hdbscan>=0.8.29
pydantic>=1.10.2
fastapi>=0.95.1
uvicorn>=0.22.0
python-dotenv>=1.0.0
anthropic>=3.0.0